Uploader: fixes and improvements
