+++ app/components/uploader/types.ts
@@ -1,17 +1,9 @@
+import { ReactChild } from 'react';
+
 export interface UploaderProps {
-  caption: string;
-  input?: any;
-  form?: any;
-  loading: boolean;
-  maxSize?: number;
-  multiple?: boolean;
-  successFile: {
-    id: string;
-    name: string;
-  };
-  setSuccessFile: (successFile) => void;
-  onDropAccepted?: (acceptedFiles) => void;
-  onDropRejected: (rejectedFiles) => void;
-  reset: (form) => void;
-  uploadFileSubmit?: () => void;
+  open?: boolean;
+  caption?: string;
+  children: ReactChild | ReactChild[];
+  onOpen: () => void;
+  onClose: () => void;
 }
